% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{map.load}
\alias{map.load}
\title{Download additional geodata of Swiss municipalities, districts, cantons and lakes}
\usage{
map.load(year, add_geodata_path = ".")
}
\arguments{
\item{year}{numeric. Available years: \code{2001-2017}.}

\item{add_geodata_path}{a character string with the path to the folder where downloaded geodata
is saved.}
}
\description{
\code{map.load} offers the possbility to download geodata of Swiss subdivisions and
   lakes for the years 2001-2017 from the official
   \href{https://github.com/zumbov2/RSwissMaps}{RSwissMaps repository} on GitHub.
}
\examples{
## Not run:
#Download Swiss geodata for 2017
map.load(2017)
## End(Not run)
}
